+++ webcomponents/monaco-editor/stencil.config.ts
@@ -1,0 +1,28 @@
+import {Config} from '@stencil/core';
+
+import {sass} from '@stencil/sass';
+import {postcss} from '@stencil/postcss';
+// @ts-ignore
+import autoprefixer from 'autoprefixer';
+
+export const config: Config = {
+  namespace: 'monaco-editor',
+  outputTargets: [
+    {
+      type: 'dist'
+    },
+    {
+      type: 'www',
+      serviceWorker: null
+    }
+  ],
+  plugins: [
+    sass(),
+    postcss({
+      plugins: [autoprefixer()]
+    })
+  ],
+  devServer: {
+    openBrowser: false
+  }
+};