+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - color
 - inline-editor
 
+### Web Components: New
+
+- monaco-editor: v1.0.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/main/webcomponents/monaco-editor/CHANGELOG.md))
+
 ### Others
 
 - utils: v4.10.0 ([CHANGELOG](https://github.com/deckgo/deckdeckgo/blob/main/utils/utils/CHANGELOG.md))